Tons of fixes for insteon:
Fix the group used to initially link new insteon devices.
Write stub handlers for all the different PLM calls to avoid a stall if
a funny message comes in.
Add a sigusr1 handler to switch into debug mode for logging
Add the ability to dump the entire ALDB from the PLM
Lots of minor fiddles and bugfixes to improve reliability.