When a 6-button keypad has button "a" hit, it sends out a message to group 3.  Previously, we were only linked as a responder on group 1. Fix:
Add defgroups to insteon.db file, listing the default groups for things
like 6 button keypads.
When the collector gets the initial aldb dump from the device, scan it for
un-linked groups, and add ourselves to them, in case of manual linking to
keypads and etc.
When doing the initial insteon_discover, use the defgroups and add
ourselves to all the groups that exist on the device normally. (ie,
keypads)

Add devcat and subcat to the insteon_devdata_t, for the discovery tool.
Unused in the collector at this time, because all-link data is just
ignored in the collector.  (collector has no use for this data)