Add jsoncgicoll, which will be used in the soon-to-be-committed code for the web interface.
Add the cfeed command.  This is a special type of feed used for switches,
which instantantly sends an upd when a swich changes.  (used in the
jsoncgicoll)

Add three new commands for an alarm subsystem.  setalarm dumpalarms
listenalarms.  These commands allow scripts and whatnot to register
arbitrary alarms with gnhast, and clients can then listen on the alarm
channel for updates, or simply dump the list of current alarms.  Alarms
are simply a user-generated uid, a severity, and a blob of text.