Add new handlers, still in alpha, for determining water usage of an irrigationcaddy system, and automatically watering based on moisture sensor readings.
Fix a bug in the handler code having to do with how int64_t values were
dealt with.  Additionally, add a new handler type, "onchange" which can be
set in the spamhandler conf argument (devices.conf only for now) which
fires the handler whenever the value changes.

Fix a bug in the scaling, where the default scale wasn't getting set
properly.