Add the astrocoll.  A collector that monitors the sun and potentially the moon.  Right now, only supports the sunrise-sunset API.  In the future, will be extended to other APIs, as well as internal calculations and monitoring of outdoor LUX devices.
Creates a sunrise device, and populates it with the sunrise/sunset current
state.  This allows for a single point of reference for the outdoor
sunrise, and not having to muck around with LUX values in every different
script to figure out what time to turn the lights on at night.